Factorized some code into AbstractCallRequest.
