+++ logisland-components/logisland-processors/logisland-processor-rest/src/main/java/com/hurence/logisland/rest/processor/lookup/AsyncCallRequest.java
@@ -45,108 +45,14 @@ import java.util.stream.Collectors;
 @Tags({"rest", "record", "http", "request", "call", "server"})
 @CapabilityDescription("Execute an http request with specified verb, body and mime type. Then stock result as a Record in the specified field")
 //@ExtraDetailFile("./details/common-processors/BulkPut-Detail.rst")
-public class AsyncCallRequest extends AbstractHttpProcessor
+public class AsyncCallRequest extends AbstractCallRequest
 {
-
-    public static final PropertyDescriptor FIELD_HTTP_RESPONSE = new PropertyDescriptor.Builder()
-            .name("field.http.response")
-            .description("The name of the field to put http response")
-            .required(false)
-            .expressionLanguageSupported(false)
-            .build();
-
-    public static final PropertyDescriptor REQUEST_METHOD = new PropertyDescriptor.Builder()
-            .name("request.method")
-            .description("The HTTP VERB Request to use.")
-            .required(false)
-            .addValidator(StandardValidators.NON_EMPTY_VALIDATOR)
-            .expressionLanguageSupported(true)
-//            .defaultValue("get")
-            .build();
-
-    public static final PropertyDescriptor REQUEST_MIME_TYPE = new PropertyDescriptor.Builder()
-            .name("request.mime.type")
-            .description("The response mime type expected for the response to use in request.")
-            .required(false)
-            .addValidator(StandardValidators.NON_EMPTY_VALIDATOR)
-            .expressionLanguageSupported(true)
-//            .defaultValue("text/plain")
-            .build();
-
-    public static final PropertyDescriptor REQUEST_BODY = new PropertyDescriptor.Builder()
-            .name("request.body")
-            .description("The body to use for the request.")
-            .required(false)
-            .addValidator(StandardValidators.NON_EMPTY_VALIDATOR)
-            .expressionLanguageSupported(true)
-            .build();
-
-    public static final PropertyDescriptor INPUT_AS_BODY = new PropertyDescriptor.Builder()
-            .name("input.as.body")
-            .description("If the input record should be serialized into json and used as body of request or not.")
-            .addValidator(StandardValidators.BOOLEAN_VALIDATOR)
-            .defaultValue("false")
-            .build();
-
-    public static final AllowableValue OVERWRITE_EXISTING =
-            new AllowableValue("overwrite_existing", "overwrite existing field", "if field already exist");
-
-    public static final AllowableValue KEEP_OLD_FIELD =
-            new AllowableValue("keep_only_old_field", "keep only old field value", "keep only old field");
-
-    public static final PropertyDescriptor CONFLICT_RESOLUTION_POLICY = new PropertyDescriptor.Builder()
-            .name("conflict.resolution.policy")
-            .description("What to do when a field with the same name already exists ?")
-            .required(false)
-            .defaultValue(KEEP_OLD_FIELD.getValue())
-            .allowableValues(OVERWRITE_EXISTING, KEEP_OLD_FIELD)
-            .build();
-
-    private String responseFieldName = null;
-    private String conflictPolicy = null;
-    private boolean inputAsBody = false;
-    private RecordSerializer serializer;
     private Vertx vertx;
-
-    @Override
-    protected Collection<ValidationResult> customValidate(ValidationContext context) {
-        final List<ValidationResult> validationResults = new ArrayList<>(super.customValidate(context));
-        if (context.getPropertyValue(INPUT_AS_BODY).asBoolean() && context.getPropertyValue(REQUEST_BODY).isSet()) {
-            validationResults.add(
-                    new ValidationResult.Builder()
-                            .input(String.format("properties '%s' and '%s' are mutually exclusive so they can not be set both at the same time",
-                                    INPUT_AS_BODY.getName(), REQUEST_BODY.getName()))
-                            .valid(false)
-                            .build());
-        }
-        return validationResults;
-    }
-
-    @Override
-    public List<PropertyDescriptor> getSupportedPropertyDescriptors() {
-        List<PropertyDescriptor> props = new ArrayList<>();
-        props.add(HTTP_CLIENT_SERVICE);
-        props.add(FIELD_HTTP_RESPONSE);
-        props.add(REQUEST_MIME_TYPE);
-        props.add(REQUEST_METHOD);
-        props.add(REQUEST_BODY);
-        props.add(INPUT_AS_BODY);
-        props.add(CONFLICT_RESOLUTION_POLICY);
-        return Collections.unmodifiableList(props);
-    }
 
     @Override
     public void init(ProcessContext context) throws InitializationException {
         super.init(context);
         try {
-            if (context.getPropertyValue(FIELD_HTTP_RESPONSE).isSet()) {
-                this.responseFieldName = context.getPropertyValue(FIELD_HTTP_RESPONSE).asString();
-            }
-            this.conflictPolicy = context.getPropertyValue(CONFLICT_RESOLUTION_POLICY).asString();
-            this.inputAsBody = context.getPropertyValue(INPUT_AS_BODY).asBoolean();
-            if (inputAsBody) {
-                serializer = SerializerProvider.getSerializer(ExtendedJsonSerializer.class.getName(), null);
-            }
             vertx = Vertx.vertx();
         } catch (Exception ex) {
             throw new InitializationException(ex);
@@ -209,26 +115,5 @@ public class AsyncCallRequest extends AbstractHttpProcessor
         Maybe.<Optional<Record>, Integer>zip(responses, (opts) -> { return 0; }).blockingGet();//wait until all request are done
         stop();
         return records;
-    }
-
-    private Optional<String> calculBody(Record record, ProcessContext context) {
-        if (context.getPropertyValue(REQUEST_BODY).isSet()) {
-            return Optional.ofNullable(context.getPropertyValue(REQUEST_BODY.getName()).evaluate(record).asString());
-        }
-        return Optional.empty();
-    }
-
-    private Optional<String> calculMimTyp(Record record, ProcessContext context) {
-        if (context.getPropertyValue(REQUEST_MIME_TYPE).isSet()) {
-            return Optional.ofNullable(context.getPropertyValue(REQUEST_MIME_TYPE.getName()).evaluate(record).asString());
-        }
-        return Optional.empty();
-    }
-
-    private Optional<String> calculVerb(Record record, ProcessContext context) {
-        if (context.getPropertyValue(REQUEST_METHOD).isSet()) {
-            return Optional.ofNullable(context.getPropertyValue(REQUEST_METHOD.getName()).evaluate(record).asString());
-        }
-        return Optional.empty();
     }
 }
